backend/requirements.txt
# --- FastAPI backend ---
fastapi==0.110.1
uvicorn==0.29.0

# --- Stable Diffusion XL and related ---
torch==2.1.0
torchvision==0.16.0
diffusers==0.26.3
transformers==4.36.2
accelerate==0.25.0
safetensors==0.4.2

# --- Utility libraries ---
pillow>=10.0.0

# --- Optional ---
python-dotenv>=1.0.0